+++ app/Jobs/PurgeExpiredSchedules.php
@@ -16,7 +16,7 @@ use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 
-class SchedulePurgeExpiredSchedules implements ShouldQueue
+class PurgeExpiredSchedules implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;