+++ database/seeders/ScheduleRecurrenceDetailsSeeder.php
@@ -2,6 +2,7 @@
 
 namespace Database\Seeders;
 
+use App\Helpers\ScheduleHelpers;
 use App\Models\Schedule;
 use App\Models\ScheduleRecurrenceDetails;
 use Carbon\Carbon;
@@ -15,9 +16,10 @@ class ScheduleRecurrenceDetailsSeeder extends Seeder
    *
    * @return void
    */
-  public function run(): void {
+  public function run(): void
+  {
     $startDateTime = Carbon::now()->subHours(6);
-    $currentStartDateTime = $startDateTime;
+    $currentStartDateTime = ScheduleHelpers::roundToNearestHalfHour($startDateTime);
 
     for ($i = 0; $i < 144; $i++) {
       $schedule = Schedule::factory()->make();
@@ -31,20 +33,28 @@ class ScheduleRecurrenceDetailsSeeder extends Seeder
       // Save the schedule
       $schedule->save();
 
-      // Create recurrence details using the custom state
+      // Create recurrence details
+      $daysOfWeek = collect(['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'])
+          ->random(rand(1, 7))
+          ->values()
+          ->toArray();
+
+      // Calculate the end date 3 months in the future that falls on one of the days_of_week
+      $endDate = ScheduleHelpers::getValidEndDate(Carbon::now()->addMonths(3), $daysOfWeek);
+
       $scheduleRecurrenceDetails = ScheduleRecurrenceDetails::factory()
           ->forSchedule($schedule)
-          ->create();
-
-      // Save the recurrence details
-      $scheduleRecurrenceDetails->save();
+          ->create([
+              'days_of_week' => json_encode($daysOfWeek),
+              'end_dateTime' => $endDate,
+          ]);
 
       // Update the schedule with the recurrence details ID
       $schedule->recurrence_details_id = $scheduleRecurrenceDetails->id;
       $schedule->save();
 
       // Update the current start time for the next schedule item
-      $currentStartDateTime = $schedule->end_dateTime;
+      $currentStartDateTime = ScheduleHelpers::roundToNearestHalfHour($schedule->end_dateTime);
     }
   }
 }